refactor: #672 github_service を API/同期境界で分割する
トークン管理・同期・API クライアント・リポジトリ要約にファイル分割し、
単一ファイルの巨大化を解消する。公開 API・NewGitHubService・interfaces は不変。
